+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [#400][] Drop dependency on `terminal-table` and use `formatador` instead. ([@netzpirat][])
 - Change the current work dir to the `watchdir`. ([@netzpirat][])
+- Add file notifier to write notifications to a configured file. ([@amiel][])
 
 ### Bug fixes